+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 2.21.3 (2026-03-14)
+
+### Fix
+
+- **test**: improve OOPIF integration test reliability
+- **iframe**: resolve nested OOPIF iframes inside shadow roots
+
+## 2.21.2 (2026-03-12)
+
+### Fix
+
+- release commit
+
 ## 2.21.1 (2026-03-03)
 
 ### Fix